Fixed up provenance semantics and execution.
